Split TestSplitter into single Classes
