+++ src/SplitTestFilesByGroupsTask.php
@@ -1,0 +1,55 @@
+<?php
+declare(strict_types = 1);
+
+namespace Codeception\Task;
+
+use Symfony\Component\Finder\Finder;
+use Symfony\Component\Finder\SplFileInfo;
+
+/**
+ * Finds all test files and splits them by group.
+ * Unlike `SplitTestsByGroupsTask` does not load them into memory and not requires Codeception to be loaded.
+ *
+ * ``` php
+ * <?php
+ * $this->taskSplitTestFilesByGroups(5)
+ *    ->testsFrom('tests/unit/Acme')
+ *    ->codeceptionRoot('projects/tested')
+ *    ->groupsTo('tests/_log/paratest_')
+ *    ->run();
+ * ?>
+ * ```
+ */
+class SplitTestFilesByGroupsTask extends TestsSplitter
+{
+    public function run()
+    {
+        $files = Finder::create()
+            ->followLinks()
+            ->name('*Cept.php')
+            ->name('*Cest.php')
+            ->name('*Test.php')
+            ->name('*.feature')
+            ->path($this->testsFrom)
+            ->in($this->projectRoot ?: getcwd())
+            ->exclude($this->excludePath);
+
+        $i = 0;
+        $groups = [];
+
+        $this->printTaskInfo('Processing ' . count($files) . ' files');
+        // splitting tests by groups
+        /** @var SplFileInfo $file */
+        foreach ($files as $file) {
+            $groups[($i % $this->numGroups) + 1][] = $file->getRelativePathname();
+            $i++;
+        }
+
+        // saving group files
+        foreach ($groups as $i => $tests) {
+            $filename = $this->saveTo . $i;
+            $this->printTaskInfo("Writing $filename");
+            file_put_contents($filename, implode("\n", $tests));
+        }
+    }
+}
